+++ tasks/tests/tasks_tests/test_fixture_module/my_config2.py
@@ -1,0 +1,1 @@
+print("In my_config 2")