+++ tasks/tests/tasks_tests/test_fixture_module/tasks.py
@@ -1,0 +1,6 @@
+from robocorp.tasks import task
+
+
+@task
+def my_task() -> str:
+    return "my_task_ran"